Edited member variable names to meet UAF coding standards, also edited constructor to create missile object at the top of the screen and a random location on the horizontal axis.Also added update and draw functions.
To be done:
Create a function to create a path to the buildings.

Note:
We decided to stick to the original game and use circular particles as missiles instead of using images as used while designing the level sketch.

This is my code and any questions or suggestions are to be directed to Arsh